Create A-Star-Navigation
This file will require an initial position of the car (will use an overhead camera for localization feature instead of hard coded at (1,1)). On the GUI, a destination input is required. The GUI displays the route the car will take to get to its destination. Note that the streets are one way only.