Aruco Positioning Updated To GUI
ASTAR algorithm is used to get shortest path from hardcoded start to hardcoded goal, it will give a list of turns to perform at each intersection; "left", "right", "straight". The list needs to be sent to the firebase for the car to read at the correct time. The time we send the next instruction to the firebase will depend on the cars positioning (open Cv color detect) relative to the oncoming intersection (Aruco marker position). Color detect is not implemented yet and will be on the next update.